Rename cubature_filter library to nonlinear_filtering
